KeyValue structure changed: is not a choice anymore
