+++ CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+
+---
+## [0.7.7] - 2022-07-28
+### Added
+### Changed
+### Fixed
+- KMIP KeyValue structure has changed in KMS from version 2.0.5
+### Removed
+---
+
