+++ src/main/java/com/cosmian/rest/kmip/data_structures/KeyValue.java
@@ -1,6 +1,11 @@
 package com.cosmian.rest.kmip.data_structures;
 
-import com.cosmian.rest.kmip.json.KmipChoice2;
+import java.util.Objects;
+import java.util.Optional;
+
+import com.cosmian.rest.kmip.json.KmipStruct;
+import com.cosmian.rest.kmip.types.Attributes;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 /**
  * The Key Value is used only inside a Key Block and is either a Byte String or a: • The Key Value structure contains
@@ -10,25 +15,67 @@ import com.cosmian.rest.kmip.json.KmipChoice2;
  * encapsulated with (and possibly wrapped with) the key material itself. • The Key Value Byte String is either the
  * wrapped TTLV-encoded Key Value structure, or the wrapped un-encoded value of the Byte String Key Material field.
  */
-public class KeyValue extends KmipChoice2<PlainTextKeyValue, byte[]> {
+public class KeyValue implements KmipStruct {
 
-    public KeyValue(Object value) {
-        super(value);
+    @JsonProperty(value = "KeyMaterial")
+    private KeyMaterial keyMaterial;
+
+    @JsonProperty(value = "Attributes")
+    private Optional<Attributes> attributes;
+
+    public KeyValue() {
+    }
+
+    public KeyValue(KeyMaterial keyMaterial, Optional<Attributes> attributes) {
+        this.keyMaterial = keyMaterial;
+        this.attributes = attributes;
+    }
+
+    public KeyMaterial getKeyMaterial() {
+        return this.keyMaterial;
+    }
+
+    public void setKeyMaterial(KeyMaterial keyMaterial) {
+        this.keyMaterial = keyMaterial;
+    }
+
+    public Optional<Attributes> getAttributes() {
+        return this.attributes;
+    }
+
+    public void setAttributes(Optional<Attributes> attributes) {
+        this.attributes = attributes;
+    }
+
+    public KeyValue keyMaterial(KeyMaterial keyMaterial) {
+        setKeyMaterial(keyMaterial);
+        return this;
+    }
+
+    public KeyValue attributes(Optional<Attributes> attributes) {
+        setAttributes(attributes);
+        return this;
     }
 
     @Override
     public boolean equals(Object o) {
-        return super.equals(o);
+        if (o == this)
+            return true;
+        if (!(o instanceof KeyValue)) {
+            return false;
+        }
+        KeyValue keyValue = (KeyValue) o;
+        return Objects.equals(keyMaterial, keyValue.keyMaterial) && Objects.equals(attributes, keyValue.attributes);
     }
 
     @Override
     public int hashCode() {
-        return super.hashCode();
+        return Objects.hash(keyMaterial, attributes);
     }
 
     @Override
     public String toString() {
-        return super.toString();
+        return "{" + " keyMaterial='" + getKeyMaterial() + "'" + ", attributes='" + getAttributes() + "'" + "}";
     }
 
 }
